Move some query building out of Query.ex
